refactor: 패키지 분리 및 변수명 수정
