+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -5,15 +5,15 @@ import nextstep.jwp.dto.LoginRequest;
 import nextstep.jwp.exception.UnauthorizedException;
 import nextstep.jwp.http.HttpCookie;
 import nextstep.jwp.http.QueryStringConverter;
-import org.apache.http.Request;
-import org.apache.http.Response;
+import org.apache.coyote.support.Request;
+import org.apache.coyote.support.Response;
 import nextstep.jwp.model.User;
 import nextstep.jwp.support.Resource;
 import nextstep.jwp.support.View;
 import nextstep.jwp.http.Session;
 import nextstep.jwp.http.SessionManager;
-import org.apache.http.HttpHeader;
-import org.apache.http.HttpStatus;
+import org.apache.coyote.HttpHeader;
+import org.apache.coyote.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
