+++ tomcat/src/main/java/nextstep/jwp/controller/ResourceController.java
@@ -1,9 +1,9 @@
 package nextstep.jwp.controller;
 
-import org.apache.http.Request;
-import org.apache.http.Response;
+import org.apache.coyote.support.Request;
+import org.apache.coyote.support.Response;
 import nextstep.jwp.support.Resource;
-import org.apache.http.HttpHeader;
+import org.apache.coyote.HttpHeader;
 
 public class ResourceController extends AbstractController {
