+++ tomcat/src/main/java/nextstep/jwp/interceptor/Interceptor.java
@@ -1,7 +1,7 @@
 package nextstep.jwp.interceptor;
 
-import org.apache.http.Request;
-import org.apache.http.Response;
+import org.apache.coyote.support.Request;
+import org.apache.coyote.support.Response;
 
 public interface Interceptor {
