+++ tomcat/src/main/java/org/apache/catalina/core/Servlet.java
@@ -1,9 +1,11 @@
-package nextstep.jwp.controller;
+package org.apache.catalina.core;
 
 import org.apache.coyote.support.Request;
 import org.apache.coyote.support.Response;
 
-public interface Controller {
+public interface Servlet {
+
+    void init();
 
     void service(Request request, Response response);
 }
