+++ tomcat/src/main/java/org/apache/coyote/Headers.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.coyote;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -7,25 +7,25 @@ public class Headers {
 
     private static final String DEFAULT_CONTENT_LENGTH = "0";
 
-    private final Map<HttpHeader, String> mapping = new LinkedHashMap<>();
+    private final Map<HttpHeader, String> values = new LinkedHashMap<>();
 
     public Headers() {
-        mapping.put(HttpHeader.CONTENT_TYPE, HttpMime.DEFAULT.getValue());
-        mapping.put(HttpHeader.CONTENT_LENGTH, DEFAULT_CONTENT_LENGTH);
+        values.put(HttpHeader.CONTENT_TYPE, HttpMime.DEFAULT.getValue());
+        values.put(HttpHeader.CONTENT_LENGTH, DEFAULT_CONTENT_LENGTH);
     }
 
     public void put(final HttpHeader key, final String value) {
-        mapping.put(key, value);
+        this.values.put(key, value);
     }
 
     public String parse() {
         final StringBuilder builder = new StringBuilder();
-        mapping.forEach((key, value) ->
+        values.forEach((key, value) ->
                 builder.append(String.format("%s: %s \r%n", key.getValue(), value)));
         return builder.toString();
     }
 
     public String find(final HttpHeader httpHeader) {
-        return mapping.get(httpHeader);
+        return values.get(httpHeader);
     }
 }
