+++ tomcat/src/main/java/org/apache/coyote/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.coyote;
 
 public enum HttpStatus {
