+++ tomcat/src/main/java/org/apache/coyote/support/RequestParser.java
@@ -1,6 +1,9 @@
-package org.apache.http;
+package org.apache.coyote.support;
 
 import nextstep.jwp.exception.FileAccessException;
+import org.apache.coyote.HttpHeader;
+import org.apache.coyote.Headers;
+import org.apache.coyote.HttpMethod;
 
 import java.io.BufferedReader;
 import java.io.IOException;
