--- tomcat/src/main/java/org/apache/http/Servlet.java
@@ -1,8 +1,0 @@
-package org.apache.http;
-
-public interface Servlet {
-
-    void init();
-
-    void service(Request request, Response response);
-}
